+++ backend/connectors/nmose/source.py
@@ -1,4 +1,4 @@
-import os
+from typing import List, Dict, Any
 
 from shapely import wkt
 from backend.connectors import NM_STATE_BOUNDING_POLYGON
@@ -20,12 +20,12 @@ class NMOSEPODSiteSource(BaseSiteSource):
     """
 
     transformer_klass = NMOSEPODSiteTransformer
-    chunk_size = 5000
+    chunk_size: int = 5000
     bounding_polygon = NM_STATE_BOUNDING_POLYGON
 
-    def get_records(self, *args, **kw) -> dict:
+    def get_records(self, *args, **kw) -> List[Dict]:
         config = self.config
-        params = {}
+        params: Dict[str, Any] = {}
         # if config.has_bounds():
         #     bbox = config.bbox_bounding_points()
         #     params["bBox"] = ",".join([str(b) for b in bbox])
@@ -37,7 +37,9 @@ class NMOSEPODSiteSource(BaseSiteSource):
         # if config.end_date:
         #     params["endDt"] = config.end_dt.date().isoformat()
 
-        url = "https://services2.arcgis.com/qXZbWTdPDbTjl7Dy/arcgis/rest/services/OSE_PODs/FeatureServer/0/query"
+        url: str = (
+            "https://services2.arcgis.com/qXZbWTdPDbTjl7Dy/arcgis/rest/services/OSE_PODs/FeatureServer/0/query"
+        )
 
         params["where"] = (
             "pod_status = 'ACT' AND pod_basin IN ('A','B','C','CC','CD','CL','CP','CR','CT','E','FS','G','GSF','H', 'HA','HC','HS','HU','J','L','LA','LRG','LV','M','MR','NH','P','PL','PN','RA','RG','S','SB','SJ','SS','T','TU','UP','VV')"
@@ -55,11 +57,14 @@ class NMOSEPODSiteSource(BaseSiteSource):
             params["geometry"] = wkt_to_arcgis_json(wkt)
             params["geometryType"] = "esriGeometryPolygon"
 
-        records = []
+        records: List = []
         i = 1
         while 1:
             rs = self._execute_json_request(url, params, tag="features")
-            records.extend(rs)
+            if rs is None:
+                continue
+            else:
+                records.extend(rs)
             params["resultOffset"] += self.chunk_size
             if len(rs) < self.chunk_size:
                 break