+++ backend/source.py
@@ -15,19 +15,13 @@
 # ===============================================================================
 from json import JSONDecodeError
 
-import click
 import httpx
 import shapely.wkt
 from shapely import MultiPoint
-from typing import Union, List
+from typing import Union, List, Callable, Dict
 
 from backend.constants import (
-    MILLIGRAMS_PER_LITER,
     FEET,
-    METERS,
-    PARTS_PER_MILLION,
-    DTW,
-    DTW_UNITS,
     DT_MEASURED,
     PARAMETER_NAME,
     PARAMETER_UNITS,
@@ -36,7 +30,6 @@ from backend.constants import (
     LATEST,
 )
 from backend.logger import Loggable
-from backend.persister import BasePersister, CSVPersister
 from backend.record import (
     AnalyteRecord,
     AnalyteSummaryRecord,
@@ -47,7 +40,7 @@ from backend.record import (
 from backend.transformer import BaseTransformer, convert_units
 
 
-def make_site_list(site_record: list | dict) -> list | str:
+def make_site_list(site_record: list[SiteRecord] | SiteRecord) -> list | str:
     """
     Returns a list of site ids, as defined by site_record
 
@@ -67,7 +60,7 @@ def make_site_list(site_record: list | dict) -> list | str:
     return sites
 
 
-def get_terminal_record(records: list, tag: Union[str, callable], bookend: str) -> dict:
+def get_terminal_record(records: list, tag: Union[str, Callable], bookend: str) -> dict:
     """
     Returns the most recent record based on the tag
 
@@ -106,6 +99,10 @@ def get_terminal_record(records: list, tag: Union[str, callable], bookend: str) 
         return sorted(records, key=func)[0]
     elif bookend == LATEST:
         return sorted(records, key=func)[-1]
+    else:
+        raise ValueError(
+            f"Invalid bookend {bookend}. Must be either {EARLIEST} or {LATEST}"
+        )
 
 
 def get_analyte_search_param(parameter: str, mapping: dict) -> str:
@@ -178,11 +175,9 @@ class BaseSource(Loggable):
     """
 
     transformer_klass = BaseTransformer
-    config = None
 
-    def __init__(self, config=None):
+    def __init__(self):
         self.transformer = self.transformer_klass()
-        self.set_config(config)
         super().__init__()
 
     @property
@@ -191,7 +186,7 @@ class BaseSource(Loggable):
 
     def set_config(self, config):
         self.config = config
-        self.transformer.config = config
+        self.transformer.set_config(config)
 
     def check(self, *args, **kw):
         return True
@@ -205,7 +200,7 @@ class BaseSource(Loggable):
     # Methods Already Implemented
     # ==========================================================================
 
-    def _execute_text_request(self, url: str, params=None, **kw) -> str:
+    def _execute_text_request(self, url: str, params: dict | None = None, **kw) -> str:
         """
         Executes a get request to the provided url and returns the text response.
 
@@ -235,8 +230,8 @@ class BaseSource(Loggable):
             return ""
 
     def _execute_json_request(
-        self, url: str, params: dict = None, tag: str = None, **kw
-    ) -> dict:
+        self, url: str, params: dict | None = None, tag: str | None = None, **kw
+    ) -> dict | None:
         """
         Executes a get request to the provided url and returns the json response.
 
@@ -256,7 +251,6 @@ class BaseSource(Loggable):
         dict
             the json response
         """
-        # print(url)
         resp = httpx.get(url, params=params, **kw)
         if tag is None:
             tag = "data"
@@ -269,18 +263,18 @@ class BaseSource(Loggable):
                 return obj
             except JSONDecodeError:
                 self.warn(f"service responded but with no data. \n{resp.text}")
-                return []
+                return None
         else:
             self.warn(f"service responded with status {resp.status_code}")
             self.warn(f"service responded with text {resp.text}")
             self.warn(f"service at url:  {resp.url}")
-            return []
+            return None
 
     # ==========================================================================
     # Methods Implemented in BaseSiteSource and BaseParameterSource
     # ==========================================================================
 
-    def read(self, *args, **kw) -> list:
+    def read(self, *args, **kw) -> list | None:
         """
         Returns the records. Implemented in BaseSiteSource and BaseAnalyteSource
         """
@@ -290,7 +284,7 @@ class BaseSource(Loggable):
     # Methods That Need to be Implemented For Each Source
     # ==========================================================================
 
-    def get_records(self, *args, **kw) -> dict:
+    def get_records(self, *args, **kw) -> List[Dict]:
         """
         Returns records as a dictionary, where the keys are site ids and
         the values are site or parameter records.
@@ -438,7 +432,7 @@ class BaseSiteSource(BaseSource):
 
         return True
 
-    def read(self, *args, **kw) -> List[SiteRecord]:
+    def read(self, *args, **kw) -> List[SiteRecord] | None:
         """
         Returns a list of transformed site records.
         Calls self.get_records, which needs to be implemented for each source
@@ -455,6 +449,7 @@ class BaseSiteSource(BaseSource):
             return self._transform_sites(records)
         else:
             self.warn("No site records returned")
+            return None
 
     def _transform_sites(self, records: list) -> List[SiteRecord]:
         """
@@ -480,7 +475,7 @@ class BaseSiteSource(BaseSource):
         self.log(f"processed nrecords={len(transformed_records)}")
         return transformed_records
 
-    def chunks(self, records: list, chunk_size: int = None) -> list:
+    def chunks(self, records: list, chunk_size: int | None = None) -> list:
         """
         Returns a list of records split into lists of size chunk_size. If
         chunk_size less than 1 then the records are not split
@@ -614,13 +609,20 @@ class BaseParameterSource(BaseSource):
         return self._extract_terminal_record(records, bookend=LATEST)
 
     def read(
-        self, site_record: SiteRecord, use_summarize: bool, start_ind: int, end_ind: int
-    ) -> List[
-        AnalyteRecord
-        | AnalyteSummaryRecord
-        | WaterLevelRecord
-        | WaterLevelSummaryRecord
-    ]:
+        self,
+        site_record: SiteRecord | list,
+        use_summarize: bool,
+        start_ind: int,
+        end_ind: int,
+    ) -> (
+        List[
+            AnalyteRecord
+            | AnalyteSummaryRecord
+            | WaterLevelRecord
+            | WaterLevelSummaryRecord
+        ]
+        | None
+    ):
         """
         Returns a list of transformed parameter records. Transformed parameter records
         are standardized so that all of the records have the same format. They are
@@ -773,6 +775,7 @@ class BaseParameterSource(BaseSource):
 
             name = ",".join(names)
             self.warn(f"{name}: No records found")
+            return None
 
     # ==========================================================================
     # Methods Implemented in BaseAnalyteSource and BaseWaterLevelSource
@@ -821,7 +824,7 @@ class BaseParameterSource(BaseSource):
     # Methods That Need to be Implemented For Each Source
     # ==========================================================================
 
-    def _extract_site_records(self, records: dict, site_record: dict) -> list:
+    def _extract_site_records(self, records: list[dict], site_record) -> list:
         """
         Returns all records for a single site as a list of records (which are dictionaries).