+++ frontend/cli.py
@@ -130,6 +130,11 @@ SPATIAL_OPTIONS = [
         default="",
         help="New Mexico county name",
     ),
+    click.option(
+        "--wkt",
+        default="",
+        help="Well known text (WKT) representation of a geometry. For example, 'POLYGON((x1 y1, x2 y2, x3 y3, x1 y1))'",
+    ),
 ]
 DEBUG_OPTIONS = [
     click.option(
@@ -210,7 +215,7 @@ def add_options(options):
 
 @cli.command()
 @click.argument(
-    "weave",
+    "parameter",
     type=click.Choice(PARAMETER_OPTIONS, case_sensitive=False),
     required=True,
 )
@@ -221,12 +226,13 @@ def add_options(options):
 @add_options(ALL_SOURCE_OPTIONS)
 @add_options(DEBUG_OPTIONS)
 def weave(
-    weave,
+    parameter,
     output,
     output_dir,
     start_date,
     end_date,
     bbox,
+    wkt,
     county,
     no_bernco,
     no_bor,
@@ -247,9 +253,8 @@ def weave(
     """
     Get parameter timeseries or summary data
     """
-    parameter = weave
     # instantiate config and set up parameter
-    config = setup_config(f"{parameter}", bbox, county, site_limit, dry)
+    config = setup_config(parameter, bbox, wkt, county, site_limit, dry)
     config.parameter = parameter
 
     # output type
@@ -289,16 +294,16 @@ def weave(
     # setup logging here so that the path can be set to config.output_path
     setup_logging(path=config.output_path)
 
+    config.report()
     if not dry:
-        config.report()
         # prompt user to continue
         if not click.confirm("Do you want to continue?", default=True):
             return
-
-    if parameter.lower() == "waterlevels":
-        unify_waterlevels(config)
-    else:
-        unify_analytes(config)
+        if parameter.lower() == "waterlevels":
+            unify_waterlevels(config)
+        else:
+            unify_analytes(config)
+    return config
 
 
 @cli.command()
@@ -308,6 +313,7 @@ def weave(
 @add_options(DEBUG_OPTIONS)
 def wells(
     bbox,
+    wkt,
     county,
     output_dir,
     no_bernco,
@@ -330,7 +336,7 @@ def wells(
     Get locations
     """
 
-    config = setup_config("sites", bbox, county, site_limit, dry)
+    config = setup_config("sites", bbox, wkt, county, site_limit, dry)
     config_agencies = [
         "bernco",
         "bor",
@@ -370,7 +376,7 @@ def wells(
     required=True,
 )
 @add_options(SPATIAL_OPTIONS)
-def sources(sources, bbox, county):
+def sources(sources, bbox, wkt, county):
     """
     List available sources
     """
@@ -381,6 +387,8 @@ def sources(sources, bbox, county):
         config.county = county
     elif bbox:
         config.bbox = bbox
+    elif wkt:
+        config.wkt = wkt
 
     parameter = sources
     config.parameter = parameter
@@ -394,7 +402,7 @@ def sources(sources, bbox, county):
         click.echo(s)
 
 
-def setup_config(tag, bbox, county, site_limit, dry):
+def setup_config(tag, bbox, wkt, county, site_limit, dry):
     config = Config()
     if county:
         click.echo(f"Getting {tag} for county {county}")
@@ -403,8 +411,14 @@ def setup_config(tag, bbox, county, site_limit, dry):
         click.echo(f"Getting {tag} for bounding box {bbox}")
         # bbox = -105.396826 36.219290, -106.024162 35.384307
         config.bbox = bbox
+    elif wkt:
+        click.echo(f"Getting {tag} for WKT {wkt}")
+        config.wkt = wkt
 
-    config.site_limit = site_limit
+    if site_limit:
+        config.site_limit = int(site_limit)
+    else:
+        config.site_limit = None
     config.dry = dry
 
     return config