+++ tests/test_sources/test_nmose_isc_seven_rivers.py
@@ -1,0 +1,16 @@
+from backend.constants import WATERLEVELS, CALCIUM, FEET, MILLIGRAMS_PER_LITER
+from tests.test_sources import BaseSourceTestClass
+
+
+class TestNMOSEISCSevenRiversWaterlevels(BaseSourceTestClass):
+
+    parameter = WATERLEVELS
+    units = FEET
+    agency = "nmose_isc_seven_rivers"
+
+
+class TestNMOSEISCSevenRiversAnalyte(BaseSourceTestClass):
+
+    parameter = CALCIUM
+    units = MILLIGRAMS_PER_LITER
+    agency = "nmose_isc_seven_rivers"